Convert to shifting instrument/track with camera and light attached
Targets are now static in space, and the perspective shifts, pushing
the instrument and track through them.  Time is consequently tied to
the instrument's z-axis position, which should simplify development
down the line, as compared to coordinating target animations and their
listening timing.

Also cleaned up some inconsistencies in the view interfaces.  Views
have a threeNode property, initialized when the object is created, that
may be added as needed to other nodes or the world.

This should probably be refactored out in the future into a ThreeView
base class or something.